Use direct channel to workers (#169)
* Make worker receive work from a channel

It is a simplification of code. It will not register itself on the
workAcquiring channel anymore, and instead all workers on a flow will
share the same incoming channel, and will get work from it.

* Make Uploader send work to workers through a simpler channel

The previous implementation has some simplicities that avoided using an
extra atomic counter I used on this new implementation. On the other
hand, it is more complex. So I'm going for the simplicity here.

* Adapt App package to deal with next uploader->worker flow

* Make the shutdown of Uploader/Worker more elegant

* Remove unnecessary atomic counter

I didn't notice the lock being held was protecting all the Enqueue
function. So it is safe to assume that this functions will not be paused
mid-flight and shutdown changed, as the lock is protecting it.